Added Gabe's Map
Ignore the fact that it's called Map1 (we can rename that later). I just finished adding the collision boxes and platforms. I left out the boxes for where the coins would be, mostly because I think that would be more of a sprint 3 objective. I want us to talk more about how that'll work later on. Let me know if there are any comments.